Refactored MainActivity in to MainPageActivity
