+++ app/src/androidTest/java/com/example/team29project/InputFragmentTest.java
@@ -19,8 +19,6 @@ import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.LargeTest;
 
 
-import com.example.team29project.View.MainActivity;
-
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
