+++ app/src/androidTest/java/com/example/team29project/MainActivityTest.java
@@ -16,7 +16,6 @@ import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.LargeTest;
 
 import com.example.team29project.Model.Item;
-import com.example.team29project.View.MainActivity;
 
 import org.junit.Rule;
 import org.junit.Test;
