+++ app/src/main/AndroidManifest.xml
@@ -5,6 +5,7 @@
     <uses-feature
         android:name="android.hardware.camera"
         android:required="true" />
+
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.CAMERA" />
@@ -21,29 +22,30 @@
         android:theme="@style/Theme.Team29project"
         tools:targetApi="31">
         <activity
-            android:name=".View.ItemViewActivity"
-            android:exported="false" />
-
-        <activity
-            android:name=".View.CustomCameraActivity"
-            android:exported="false" />
-
-        <activity
             android:name=".View.MainActivity"
-            android:exported="true">
+            android:exported="false">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".View.ItemViewActivity"
+            android:exported="false" />
+        <activity
+            android:name=".View.CustomCameraActivity"
+            android:exported="false" />
+        <activity
+            android:name=".View.MainPageActivity"
+            android:exported="false"></activity>
 
         <meta-data
             android:name="com.google.android.gms.vision.DEPENDENCIES"
             android:value="barcode" />
         <meta-data
             android:name="com.google.mlkit.vision.DEPENDENCIES"
-            android:value="barcode_ui"/>
+            android:value="barcode_ui" />
     </application>
 
 </manifest>
