+++ app/src/main/java/com/example/team29project/Model/User.java
@@ -1,0 +1,12 @@
+package com.example.team29project.Model;
+
+import java.util.ArrayList;
+
+public class User {
+    private String uniqueName;
+    private String password;
+
+    public User(){
+    }
+
+}
